+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Removed
 
 - Removed deprecated `description_source` property from Mastodon API Instance.
+- Stop accepting legacy identity proofs.
 
 ## [2.12.0] - 2024-02-26